Diplom_work/requirements.txt
requests>=2.31.0
allure-pytest>=2.13.2
pytest>=7.4.0
selenium>=4.10.0
pytest-html>=3.2.0
webdriver-manager>=3.8.6